requirements.txt
Pillow
requests
fastapi
uvicorn
gunicorn
